优化解绑邮箱手机不能发送短信问题

+++ RearEnd/SmallProfitMall/src/main/java/cn/itcast/service/impl/FaceRecognitionServiceImpl.java
@@ -189,7 +189,7 @@ public class FaceRecognitionServiceImpl implements FaceRecognitionService {
                 String score="score";
                 JSONObject result = videoValidation.getJSONObject("result");
                 //活体分值 小于0.88 返回不成功
-                if (result.getInt(score)<CREDIT){
+                if (result.getDouble(score)<CREDIT){
                     JSONObject videoError = new JSONObject();
                     videoError.put("error_code",ERROR_CODE);
                     return videoError.toString();
